+++ CHANGELOG.md
@@ -27,6 +27,9 @@
 
 ### Changed
 - Updated the view of the features tab to match the style of the abilities tab (#153).
+  - The kit section will only be visible if the actor's class allows kits or there's kits on the actor.
+  - Made the kit swap dialog more clear on why you need to swap kits.
+- Added an embed for kits used for viewing on the actor sheet, viewing in play mode on the kit sheet, and posting to chat. (#239)
 
 ### Fixed
 - Added zero-width spaces after "/" in names to ensure they linebreak properly (#241).